pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.133.1",
    "groq>=0.37.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "sentencepiece>=0.2.1",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
